requirements.txt
asgiref==3.6.0
Django==4.1.5
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2022.7
PyYAML==6.0
sqlparse==0.4.3
djangorestframework-simplejwt==5.2.2
